feat: 完成Rustlings全部练习

+++ exercises/conversions/using_as.rs
@@ -10,16 +10,15 @@
 // Execute `rustlings hint using_as` or use the `hint` watch subcommand for a
 // hint.
 
-// I AM NOT DONE
-
 fn average(values: &[f64]) -> f64 {
     let total = values.iter().sum::<f64>();
-    total / values.len()
+    // 关键修改：将 usize 类型的长度转换为 f64 类型
+    total / values.len() as f64
 }
 
 fn main() {
     let values = [3.5, 0.3, 13.0, 11.7];
-    println!("{}", average(&values));
+    println!("{}", average(&values)); // 输出：7.125
 }
 
 #[cfg(test)]
